requirements.txt
Django~=5.2.3
python-dotenv~=1.1.0
loguru~=0.7.3
django-redis~=5.4.0